+++ apps/playground/src/index.ts
@@ -2,12 +2,14 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 import { SqliteDatabaseAdapter } from "@elizaos/adapter-sqlite";
 import DirectClientInterface from "@elizaos/client-direct";
+import TelegramClientInterface from "@elizaos/client-telegram";
 import { AgentBuilder, ModelProviderName } from "@iqai/agent";
 import { createAtpPlugin } from "@iqai/plugin-atp";
-import { createBAMMPlugin } from "@iqai/plugin-bamm";
 import { createJsPlugin } from "@iqai/plugin-js";
 import { createFraxlendPlugin } from "@iqai/plugin-fraxlend";
+import createHeartbeatPlugin from "@iqai/plugin-heartbeat";
 import { createOdosPlugin } from "@iqai/plugin-odos";
+import { createWalletPlugin } from "@iqai/plugin-wallet";
 import createSequencerPlugin from "@iqai/plugin-sequencer";
 import Database from "better-sqlite3";
 import { fraxtal } from "viem/chains";
@@ -27,15 +29,58 @@ async function main() {
 	const atpPlugin = await createAtpPlugin({
 		walletPrivateKey: process.env.WALLET_PRIVATE_KEY,
 	});
+
+	// const nearPlugin = await createNearPlugin({
+	// 	accountId: process.env.NEAR_ACCOUNT_ID as string,
+	// 	accountKey: process.env.NEAR_PRIVATE_KEY as string,
+	// 	listeners: [
+	// 		{
+	// 			eventName: "run_agent",
+	// 			contractId: "amm-iqai.testnet",
+	// 			responseMethodName: "agent_response",
+	// 			handler: async (payload, { account }) => {
+	// 				const request = JSON.parse(payload.message);
+
+	// 				const balances = await account.viewFunction({
+	// 					contractId: "amm-iqai.testnet",
+	// 					methodName: "get_swap_balances",
+	// 					args: {
+	// 						token_in: request.token_in,
+	// 						token_out: request.token_out,
+	// 					},
+	// 				});
+
+	// 				const balance_in = BigInt(balances[0]);
+	// 				const balance_out = BigInt(balances[1]);
+	// 				const amount_in = BigInt(request.amount_in);
+
+	// 				const k = balance_in * balance_out;
+	// 				const new_balance_in = balance_in + amount_in;
+
+	// 				if (amount_in > 0 && new_balance_in > 0) {
+	// 					const new_balance_out = k / new_balance_in;
+	// 					const amount_out = balance_out - new_balance_out;
+	// 					return amount_out.toString();
+	// 				}
+
+	// 				throw new Error("Illegal amount");
+	// 			},
+	// 		},
+	// 	],
+	// 	networkConfig: {
+	// 		networkId: "testnet",
+	// 		nodeUrl: "https://test.rpc.fastnear.com",
+	// 	},
+	// });
+
 	const sequencerPlugin = await createSequencerPlugin();
 
-	const bammPlugin = await createBAMMPlugin({
-		walletPrivateKey: process.env.WALLET_PRIVATE_KEY,
-		chain: fraxtal,
+	const walletPlugin = await createWalletPlugin({
+		covalentApiKey: process.env.COVALENT_API_KEY,
+		walletAddress: process.env.WALLET_ADDRESS,
 	});
 
 	const jsPlugin = await createJsPlugin();
-
 	// Setup database
 	const dataDir = path.join(process.cwd(), "./data");
 	fs.mkdirSync(dataDir, { recursive: true });
@@ -46,6 +91,7 @@ async function main() {
 	const agent = new AgentBuilder()
 		.withDatabase(databaseAdapter)
 		.withClient("direct", DirectClientInterface)
+		.withClient("telegram", TelegramClientInterface)
 		.withModelProvider(
 			ModelProviderName.OPENAI,
 			process.env.OPENAI_API_KEY as string,
@@ -54,9 +100,8 @@ async function main() {
 			fraxlendPlugin,
 			odosPlugin,
 			atpPlugin,
+			// nearPlugin,
 			sequencerPlugin,
-			bammPlugin,
-			jsPlugin,
 		])
 		.withCharacter({
 			name: "BrainBot",